+++ TechnicalProcessControl.DAL/Models/Drawing.cs
@@ -28,7 +28,7 @@ namespace TechnicalProcessControl.DAL.Models
         public short? QuantityL { get; set; }
 
         public DateTime? CreateDate { get; set; }
-        public short? ParentId { get; set; }
+        public int? ParentId { get; set; }
         public string Note { get; set; }
     }
 }
